Issue #33 - Make changes to separate component
- Removed redundant files (Table1,Table2,Dashboard1,Dashboard2)
- Moved Settings to App section
